Move GetSourceFiles to Source class
